test: refactor _patch_dummy_writer
