+++ tests/contrib/pytest_bdd/test_pytest_bdd.py
@@ -10,7 +10,7 @@ from ddtrace.contrib.pytest.plugin import is_enabled
 from ddtrace.contrib.pytest_bdd.plugin import _get_step_func_args_json
 from ddtrace.ext import test
 from ddtrace.internal.ci_visibility import CIVisibility
-from tests.ci_visibility.test_encoder import _patch_dummy_writer
+from tests.ci_visibility.util import _patch_dummy_writer
 from tests.utils import DummyCIVisibilityWriter
 from tests.utils import TracerTestCase
 from tests.utils import override_env
